chore(supabase_test): move the internal suite fixtures to an internal library and add an example
getSessionData, sessionDataUserId, todoColumns, FailingHttpClient,
JsonResponseMockClient and stallUntilAborted only serve the test suites of
the supabase-flutter monorepo. They now live in
package:supabase_test/internal.dart, so the main library carries only what
an app test needs. The example directory shows the quick start as a runnable
test file.
