+++ packages/postgrest/test/test_utils.dart
@@ -10,6 +10,7 @@ library;
 import 'package:supabase_common/testing.dart';
 
 export 'package:supabase_common/testing.dart';
+export 'package:supabase_test/internal.dart';
 export 'package:supabase_test/supabase_test.dart';
 
 const apiHeaders = {
