+++ packages/supabase/test/mock_test.dart
@@ -6,6 +6,7 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:supabase/supabase.dart';
+import 'package:supabase_test/internal.dart';
 import 'package:supabase_test/supabase_test.dart';
 import 'package:test/test.dart';
