+++ packages/supabase/test/utils.dart
@@ -1,6 +1,7 @@
 import 'package:supabase/supabase.dart';
 
 export 'package:supabase_common/testing.dart';
+export 'package:supabase_test/internal.dart';
 export 'package:supabase_test/supabase_test.dart';
 
 class TestAsyncStorage extends MemoryAuthAsyncStorage {}
