+++ packages/supabase_auth/test/utils.dart
@@ -3,6 +3,7 @@ import 'package:supabase_auth/supabase_auth.dart';
 import 'package:supabase_common/testing.dart';
 
 export 'package:supabase_common/testing.dart';
+export 'package:supabase_test/internal.dart';
 export 'package:supabase_test/supabase_test.dart';
 
 /// Email of a user with unverified factor
