+++ packages/supabase_flutter/test/utils.dart
@@ -3,6 +3,7 @@ import 'package:shared_preferences_platform_interface/in_memory_shared_preferenc
 import 'package:shared_preferences_platform_interface/shared_preferences_async_platform_interface.dart';
 
 export 'package:supabase_common/testing.dart';
+export 'package:supabase_test/internal.dart';
 export 'package:supabase_test/supabase_test.dart';
 
 /// Replaces both shared_preferences APIs with empty in-memory stores.
