+++ packages/supabase_storage/test/custom_http_client.dart
@@ -3,7 +3,7 @@ import 'dart:typed_data';
 
 import 'package:http/http.dart';
 
-export 'package:supabase_test/supabase_test.dart' show FailingHttpClient;
+export 'package:supabase_test/internal.dart' show FailingHttpClient;
 
 /// Client that fails for few times when attempting to upload file
 class RetryHttpClient extends BaseClient {
